# core: speed up Status code and message parsing