@carl-mastrangelo 2016-09-08
One last thing:  The difference here is noticeable:

```
Before:
Benchmark                         (direct)  (transport)    Mode      Cnt       Score     Error  Units
TransportBenchmark.unaryCall1024      true        NETTY  sample  3205966  155710.268 ± 149.278  ns/op

After:
Benchmark                         (direct)  (transport)    Mode      Cnt       Score     Error  Units
TransportBenchmark.unaryCall1024      true        NETTY  sample  3385015  147474.794 ± 128.733  ns/op
```

I included the full numbers in the second commit message, but this better.  The 99% latency goes down by 25us!