@buchgr 2016-09-08
@carl-mastrangelo wow! very nice! 6% faster! I ll have a look at it tomorrow morning . it's almost midnight here 😴 :)