@buchgr 2016-09-13
LGTM, once remaining comments are addressed and build passes. Good work @carl-mastrangelo !